Switch to u64 weights (#28)
This has the drawback of allowing for the possibility of weight overflows, but that's highly unlikely to be a problem in practice.

If it happens (and overflow checks disabled) it might cause the cache to get confused (e.g. allowing more items or infinite eviction loops) but it cannot lead to memory unsafety.

+++ fuzz/fuzz_targets/fuzz_cache.rs
@@ -16,8 +16,8 @@ struct MyWeighter;
 struct MyLifecycle;
 
 impl Weighter<u16, (u16, u16)> for MyWeighter {
-    fn weight(&self, _key: &u16, val: &(u16, u16)) -> u32 {
-        val.1 as u32
+    fn weight(&self, _key: &u16, val: &(u16, u16)) -> u64 {
+        val.1 as u64
     }
 }
